requirements.txt
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4
requests==2.32.3
matplotlib==3.9.2
python-dotenv==1.0.0
